--- app/routes/o.$orgSlug/dashboard/form-preferences/-components/loading-form-preference.tsx
@@ -1,13 +1,0 @@
-import { Empty, EmptyMedia, EmptyTitle } from "@/components/ui/empty";
-import { Spinner } from "@/components/ui/spinner";
-
-export default function LoadingFormPreference() {
-  return (
-    <Empty>
-      <EmptyMedia>
-        <Spinner className="size-8" />
-      </EmptyMedia>
-      <EmptyTitle>Loading form preference...</EmptyTitle>
-    </Empty>
-  );
-}
