--- components/blocks/markdown-editor-with-links/plugins.tsx
@@ -1,60 +1,0 @@
-import { LINK } from "@lexical/markdown";
-import { ClickableLinkPlugin } from "@lexical/react/LexicalClickableLinkPlugin";
-import { LexicalErrorBoundary } from "@lexical/react/LexicalErrorBoundary";
-import { MarkdownShortcutPlugin } from "@lexical/react/LexicalMarkdownShortcutPlugin";
-import { RichTextPlugin } from "@lexical/react/LexicalRichTextPlugin";
-import { useState } from "react";
-import { ContentEditable } from "@/components/editor/editor-ui/content-editable";
-import { AutoLinkPlugin } from "@/components/editor/plugins/auto-link-plugin";
-import { FloatingLinkEditorPlugin } from "@/components/editor/plugins/floating-link-editor-plugin";
-import { LinkPlugin } from "@/components/editor/plugins/link-plugin";
-import { LinkToolbarPlugin } from "@/components/editor/plugins/toolbar/link-toolbar-plugin";
-import { ToolbarPlugin } from "@/components/editor/plugins/toolbar/toolbar-plugin";
-
-export function Plugins() {
-  const [floatingAnchorElem, setFloatingAnchorElem] =
-    useState<HTMLDivElement | null>(null);
-  const [isLinkEditMode, setIsLinkEditMode] = useState(false);
-
-  const onRef = (_floatingAnchorElem: HTMLDivElement) => {
-    if (_floatingAnchorElem !== null) {
-      setFloatingAnchorElem(_floatingAnchorElem);
-    }
-  };
-
-  return (
-    <div className="relative">
-      <ToolbarPlugin>
-        {() => (
-          <div className="vertical-align-middle sticky top-0 z-10 flex gap-2 overflow-auto border-b p-1">
-            <LinkToolbarPlugin setIsLinkEditMode={setIsLinkEditMode} />
-          </div>
-        )}
-      </ToolbarPlugin>
-      <div className="relative">
-        <RichTextPlugin
-          contentEditable={
-            <div className="">
-              <div className="" ref={onRef}>
-                <ContentEditable
-                  placeholder="Start typing..."
-                  className="ContentEditable__root relative block h-72 min-h-72 overflow-auto px-8 py-4 focus:outline-none"
-                />
-              </div>
-            </div>
-          }
-          ErrorBoundary={LexicalErrorBoundary}
-        />
-        <ClickableLinkPlugin />
-        <AutoLinkPlugin />
-        <LinkPlugin />
-        <FloatingLinkEditorPlugin
-          anchorElem={floatingAnchorElem}
-          isLinkEditMode={isLinkEditMode}
-          setIsLinkEditMode={setIsLinkEditMode}
-        />
-        <MarkdownShortcutPlugin transformers={[LINK]} />
-      </div>
-    </div>
-  );
-}
